Move almost all interfaces to a single file
https://coveord.atlassian.net/browse/JSUI-2725
