+++ unitTests/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValueTest.ts
@@ -3,11 +3,10 @@ import { DynamicFacetValue } from '../../../../src/ui/DynamicFacet/DynamicFacetV
 import { DynamicFacetTestUtils } from '../DynamicFacetTestUtils';
 import { DynamicFacetRangeTestUtils } from '../DynamicFacetRangeTestUtils';
 import { DynamicFacet } from '../../../../src/ui/DynamicFacet/DynamicFacet';
-import { IDynamicFacetOptions } from '../../../../src/ui/DynamicFacet/IDynamicFacet';
+import { IDynamicFacetOptions, IDynamicFacetValue } from '../../../../src/ui/DynamicFacet/IDynamicFacet';
 import { FacetValueState } from '../../../../src/rest/Facet/FacetValueState';
 import { analyticsActionCauseList } from '../../../../src/ui/Analytics/AnalyticsActionListMeta';
 import { DynamicFacetValueRenderer } from '../../../../src/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValueRenderer';
-import { IDynamicFacetValue } from '../../../../src/ui/DynamicFacet/DynamicFacetValues/IDynamicFacetValue';
 
 export function DynamicFacetValueTest() {
   describe('DynamicFacetValue', () => {
